pyproject.toml
[tool.poetry]
name = "vassouras-ii"
version = "0.1.0"
description = ""
authors = ["pauloandrade1985 <paulo.smandrade@hotmail.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.2"
sqlalchemy = "^2.0.43"
pydantic-settings = "^2.10.1"
flask-socketio = "^5.5.1"
alembic = "^1.16.5"
pytz = "^2025.2"
tzdata = "^2025.2"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
flask-login = "^0.6.3"
eventlet = "^0.40.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.10"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude =['migrations']

[tool.ruff.lint]
preview = true
select =['I','F','E','W','PL','PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = "flask --app vassouras_ii.app run --debug --port=5000 --host=0.0.0.0"
pre_test = 'task lint'
test = 'pytest -s -x --cov=vassouras_ii -vv'
post_test = 'coverage html'
